TPASPKT-478 Fix tests
Fix unit test and an E2E test in 'dpres-research-rest-api'.

file-scraper has a new ResultsMergeScraper that causes a new PREMIS
agent XML file to be created.

Unlike the other agent XML files, it also seems to be eventually
removed. This is probably because it remains unused and isn't used to
create a single PREMIS event, although this isn't confirmed.

In addition, don't assume the 'file_characteristics_extension/streams'
field exists.