TPASPKT-649 Fix handling of unrecognized files
Since TPASPKT-777, file-scraper provides MIME type also for file formats
that do not have a scraper. Therefore, it does not mean that a file is
really recognized if MIME type is available. The MIME type could be for
example "inode/x-empty" or "application/octet-stream". For recognized
files also the format_version should be available (or unapplicable).