TPAPSKT-524 Allow packaging SIP without sending it to DPS
Added possibilty to run workflow that only creates SIP, but does not
send it to DPS. This workflow can be used for example to validate
metadata of dataset. The preservation workflow can reuse the outputs
from earlier packaging workflow to avoid creating the same SIP twice.

Moreover, scheduling new workflow was simplified a bit:

* When new workflow is created, luigi is not run interactively. The
  workflow is just added to workflow database, and it is started by
  running the InitWorkflows task.

* The preservation state is not automatically set when workflow is
  created. The preservation state should be set before creating the
  workflow.