TPASPKT-1008 Fix single stream linking
siptools has a special case when only one AudioMD or VideoMD stream is
detected with the index 0: the stream metadata is applied to the file
itself rather than being considered a stream inside the file.

siptools-research did not implement this special case, causing a missing
reference in the generated METS later on. Although the issue was
discovered in AudioMD, also implement for it VideoMD since both of them
implement the same special behavior.