TPASPKT-1000 Add logrotate configuration
Add logrotate configuration to the RPM package, as is done with other
packages on RHEL7 such as Redis.

dpres-siptools-research can create very large log files due to
comprehensive logging of SSH connections when polling for finished
packages. This has exhausted the disk space on several occasions. Log
rotation also compresses the log files, which should take care of this
problem.