TPASPKT-1109 Update Fairdata download service DNS
Fairdata has updated their policy on how to access Fairdata download
service.

The download service should be referred to with the domain name
download.* in all configurations and by all services in all
environments. It should not be accessed using the IDA domain names.

The fact that the download service is hosted on an IDA server instance
does not diminish the fact that it is a distinct service and could later
be hosted independently, and the integrations should respect and reflect
that.