TPASPKT-861 Merge all workflow tasks into same workflow
All workflow tasks are now part of the packaging workflow, but it is possible
to run only part of the workflow by initializing metadata generation
workflow of validation workflow. When a workflow is initialized, only
the "target task" of the previous of workflow is updated (if a previous
workflow exists). If the previous workflow should not be continued,
for example when metadata has been edited after validation, the previous
workflow is disabled, and new workflow is created.

Note that the metadata of the same dataset is used throughout the whole
worfklow, so when the "PAS version" is created from the original Ida
dataset, the original dataset is used for packaging. This means that the
identifier of the original dataset must always be used when initializing
workflows.