TPASPKT-1269 Scrape files in workspace, not in "file cache"
The "file cache" was implemented, because before TPASPKT-734 the
metadata generation and file validation happened outside luigi
workflows, and therefore files could not be saved to the "workspace"
directory. As metadata generation and file validation are now part of
the packaging workflow, it is logical that they use the same files as
other tasks in the workflow. Moreover, scraping cached files was
problematic, because files in file cache do not have filename extension,
which scraper uses for detection/validation.

So now `GenerateMetadata` and `ValidateFiles` tasks require `GetFiles`
task has downloaded files to workspace, which requires a few changes to
the workflow:

* Metadata validation can not be done before `GetFiles`, because it
  validates metadata that will be generated by GenerateMetadata. This
  should not be problem, since all metadata used by `GetFiles` should
  always be valid.

* The files will only be downloaded when metadata generation workflow is
  started/restarted. So if dataset files are changed after metadata
  generation, metadata generation workflow must be restarted. This does
  not really make any difference, as packaging would fail anyway if user
  changes the dataset files during preservation process.

* `GetFiles` task will continue to download files, even if one of the
  dataset file missing. This functionality was previously implemented in
  ValidateFiles task.