TPASPKT-483 Remove workflow actions from CLI
Remove 'preserve', 'validate' and 'generate' actions from the CLI. The
implementation for these commands are partially implemented in
dpres-admin-rest-api; for example, validation and Metax state changes
are performed in dpres-admin-rest-api.

Since the commands at the moment are incomplete and possibly broken,
remove them completely for now. They can be reintroduced later if
necessary (eg. administrators need to handle datasets in bulk which is
inconvenient with the web UI).