TPASPKT-648 Create METS document in one task
Previously parts of METS document have been created in five different
tasks:

* CreateProvenanceInformation
* CreateDescriptiveMetadata
* CreateLogicalStructMap
* CreateStructMap
* CreateTechnicalMetadata

The METS document parts have then been combined to one file in
CreateMets task. To simplify the workflow, all functionality of the
previously mentioned tasks is merged to CreateMets. To keep the diff
understandable, this change is done in two steps. In this first step,
all of the tasks are imported to CreateMets task, which will simply
call the run methods of other tasks when it is run. In the second step
(in next commit) the code from the old tasks module is actually merged
to CreateMets, and the old modules are removed.

As The METS part creation tasks are not really used as luigi Tasks
anymore, they can then be simplified to make the merge easier:

* `requires` `output` methods are removed
* All hacks that make the tasks atomic and idempotent are removed