TPASPKT-1107 Allow files without use category
Before this commit, the use category of directory was automatically
applied to subdirectories and files in that directory. Files without use
category were not allowed, and dataset was considered invalid if the
use category was not defined to every file (or one of the parent
directories of the file).

After this commit, files without use category are allowed, but they will
not be added to logical structural map. The use category of parent
directory is omitted.