TPASPKT-1397 Remove `pas_storage_id` config param
There is a hardcoded list of allowed values in Metax and this value was
essentially the same everywhere, so keeping this configuration
parameter around doesn't seem to make sense anymore.

The "storage identifier" term used in several places in the code is also
ambiguous due to the new `storage_identifier` field in Metax V3.
Removing the configuration parameter also resolves this ambiguity.