TPASPKT-1434 Remove 'preservation_event'
Do not use the `preservation_event` field when creating the provenance
metadata in the preservation package.

While `preservation_event` field was required in the absence of the
`lifecycle_event` field, no one used `preservation_event`. Removing the
field simplifies the workflow slightly as we no longer have to ensure
either of the two is set, just `lifecycle_event`.