TPASPKT-1464 Convert CSV record separator
There are two representations of the CSV record separator:

* The "literal" method using the characters themselves as-is: `\r`, `\n`
  and `\r\n`. file-scraper uses this for both input and output.
* The enum used by Metax V3 which enforces constants in
  `file_characteristics.csv_record_separator` to refer to accepted
  values: `CR`, `LF` and `CRLF`

Metax V3 does not accept the first format, which means we will have to
convert between the two depending on whether we are interfacing with
file-scraper or Metax V3. In Metax V3, we can leave
`file_characteristics_extension` unchanged, as it can contain any
arbitrary JSON data.

The conversion process is clunky and needs to be applied in several
places. In an ideal case, it would probably be easier if Metax V3
supported the literal characters as it did with Metax V2. This might be
implemented at a later date; for now, it's more important to get CSV
metadata generation into a working condition.