TPASPKT-1378 Link bit-level files
Link bit-level and DPRES compatible files together if the linkings exist
on the Metax side (i.e. `pas_compatible_file` and`
`non_pas_compatible_file` file).

If a file with a DPRES compatibility linking exists in the dataset, both
files have to be included with the same dataset; attempting to submit a
dataset with a lone file will result in an error.

The dataset files are checked during `ValidateMetadata` task. This
happens before the PAS catalog copy of the dataset is made.