TPASPKT-1378 Permit bit-level files
Bit-level files are allowed in FDDPS as long as they're accompanied by
a source file using an accepted or recommended file format; this is
enforced by the previous commit.

To permit this, allow the file format version to be left empty for
bit-level files. In practice, this means we have to short-circuit
and skip metadata validation in several places if we are dealing with a
bit-level file linked to a DPRES compatible file.

File format detection is also optional in this scenario.