TPASPKT-1378 Add binary test file
Add binary "test file" which is just 2 KB of pseudo-random data piped into a
to a file. This will be the stand-in for a bit-level file which is not
recognized by file-scraper.

The file was created with the following command (using Python 3.9.21-1
on AlmaLinux 9):

```
python3 -c "import random; import sys; random.seed(42); sys.stdout.buffer.write(random.randbytes(2048))" > tests/data/sample_files/random_data.bin
```