TPASPKT-1378 Validate PAS file pair grades
Check two cases where PAS compatible file pairs need certain
file-scraper grades:

* 'bit-level-with-recommended' file must be linked to a PAS compatible
  file in Metax
* PAS compatible file in Metax must have the 'recommended' grade

Note that there is also the 'bit-level' grade in which case the file is
fine as-is and does not need to be linked to another file in Metax.

There is only one file yet in file-scraper test file corpus with the
'bit-level-with-recommended'. However, this file cannot yet be
submitted due to a bug in dpres-siptools-ng in technical metadata
generation; this will be resolved in a different ticket.