TPASPKT-1415 Add API to retrieve file errors
Add simple API endpoint in research-rest-api to retrieve file errors.
dpres-admin-rest-api cannot access the file error database directly,
hence the need for the endpoint.

The endpoint is deliberately very simple; the assumption is that a more
elaborate user-facing API (eg. with input validation, possible
pagination, etc.) will be implemented in admin-rest-api. That API and
the frontend implementation will be implemented in TPASPKT-1416.