Add dataset link to file link error
Set `is_dataset_error` on the exception raised when dataset files are
incomplete and contain incomplete PAS file links.

This is now needed because we want to store the dataset identifier to
the file error database when it's relevant to the error. At the moment
this is the only exception where that is the case.