TPASPKT-1499 Use MongoBox
MongoBox launches an ephemeral in-memory MongoDB database for each
pytest session. This is already in use in several DPRES projects
including `upload-rest-api` and `preservation`.

Previously we monkeypatched every place where a MongoDB connection might
be made, which is error-prone and means that unpatched calls would just
use the MongoDB process running outside the test suite in the default
port, potentially breaking tests and causing the actual MongoDB
database's data to be modified.