TPASPKT-928 Add function to build bulk file exc
Add context manager to build bulk file exceptions. The context manager
is supposed to be used like this:

```
with file_error_collector() as collect:
    for file in dataset_files:
        if not validate_file(file):
            collect(InvalidDatasetFileError("File is not valid"))
```

This is done because the exception needs to be raised in two scenarios:

* we ran the workflow task to completion, but some errors were found
* we tried running the workflow task, but too many errors were found and
  the workflow task should be halted prematurely

Without a context manager, two blocks of code would have to be
implemented for both cases in each workflow task. A context manager
allows us to encapsulate both cases into a single `with` block.