TPASPKT-1597 Divide dataset.py to smaller modules
The Dataset class has been divided into several classes:

* Dataset
   * Manages the dataset entry in Metax.
   * Stores error messages related to the dataset.
* Workflow
   * Maps workflows to their target tasks.
   * Determines when specific actions are permitted.
   * Creates workspaces for workflows.
   * Initializes workflows.
   * Manages workflow status. For example, whether the dataset metadata
     has been confirmed or whether the dataset has been proposed for
     preservation.
* Workspace
   * Maps workflows to their associated workspace directories.
   * Is used by both Workflow and WorkflowTasks.

The previous method `dataset.find_datasets` has been replaced by
`workflow.find_workflows`, because workflows are now the primary
entities we want to locate. The old method would have been used only in
tests, so there was no reason to keep it.

WorkflowTasks are aware of the Dataset they operate on, but they are not
aware of the Workflow that manages those tasks. As a result,
WorkflowTasks can no longer directly "disable" a workflow when a target
task completes or when an error is detected in the dataset. A new
mechanism for stopping workflows will be introduced in upcoming commits.
From this point forward, the term "disabled" refers exclusively to
workflows that have been manually disabled by administrators.

Note that although some code has been moved between modules, existing
tests remain in their original test modules to make it easier to see how
functionality has changed. The tests will be reorganized in following
commits.